+++ NEWS.md
@@ -6,6 +6,7 @@
 ## new features
 * dataAugmentation / imagesToKerasInput: new argument "subset" (for subsetting function input)
 * imagesToKerasInput: new argument "max" (to specify maximum value of image color range)
+* resizeImages: new argument "filter", to specify filter algorithm for resizing.
 
 ## other changes
 * u_net uses layer_conv_2d_transpose instead of layer_upsampling_2d